Refactor tests to use namespace and shared stubs
