+++ tests/CookiesTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-use PHPUnit\Framework\TestCase;
-use Lotgd\Cookies;
+namespace Lotgd\Tests;
 
-require_once __DIR__ . '/../config/constants.php';
+use Lotgd\Cookies;
+use PHPUnit\Framework\TestCase;
 
 final class CookiesTest extends TestCase
 {
