+++ tests/NamesTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-use PHPUnit\Framework\TestCase;
-use Lotgd\Names;
+namespace Lotgd\Tests;
 
-require_once __DIR__ . '/../config/constants.php';
+use Lotgd\Names;
+use PHPUnit\Framework\TestCase;
 
 final class NamesTest extends TestCase
 {
