+++ tests/NavigationItemTest.php
@@ -2,58 +2,15 @@
 
 declare(strict_types=1);
 
-namespace {
-    use PHPUnit\Framework\TestCase;
-    use Lotgd\Nav\NavigationItem;
-    use Lotgd\Output;
-    use Lotgd\Template;
-    use Lotgd\Nav;
+namespace Lotgd\Tests;
 
-    require_once __DIR__ . '/../config/constants.php';
-    if (!function_exists('modulehook')) {
-        function modulehook($name, $data = [], $allowinactive = false, $only = false)
-        {
-            return $data;
-        }
-    }
-    if (!function_exists('translate')) {
-        function translate($t, $ns = false)
-        {
-            return $t;
-        }
-    }
-    if (!function_exists('translate_inline')) {
-        function translate_inline($t, $ns = false)
-        {
-            return $t;
-        }
-    }
-    if (!function_exists('tlbutton_pop')) {
-        function tlbutton_pop()
-        {
-            return '';
-        }
-    }
-    if (!function_exists('tlschema')) {
-        function tlschema($schema = false)
-        {
-        }
-    }
-    if (!function_exists('popup')) {
-        function popup(string $page, string $size = '550x300')
-        {
-            return '';
-        }
-    }
-    if (!function_exists('appoencode')) {
-        function appoencode($data, $priv = false)
-        {
-            global $output;
-            return $output->appoencode($data, $priv);
-        }
-    }
+use Lotgd\Nav;
+use Lotgd\Nav\NavigationItem;
+use Lotgd\Output;
+use Lotgd\Template;
+use PHPUnit\Framework\TestCase;
 
-    final class NavigationItemTest extends TestCase
+final class NavigationItemTest extends TestCase
     {
         protected function setUp(): void
         {
@@ -80,4 +37,3 @@ namespace {
             $this->assertStringContainsString('navhi', $html);
         }
     }
-}
