+++ tests/UtilityTest.php
@@ -2,22 +2,15 @@
 
 declare(strict_types=1);
 
-use PHPUnit\Framework\TestCase;
-use Lotgd\EmailValidator;
-use Lotgd\SafeEscape;
-use Lotgd\Stripslashes;
+namespace Lotgd\Tests;
+
 use Lotgd\Dhms;
-use Lotgd\Sanitize;
+use Lotgd\EmailValidator;
 use Lotgd\Output;
-
-require_once __DIR__ . '/../config/constants.php';
-
-if (!function_exists('translate_inline')) {
-    function translate_inline($text, $ns = false)
-    {
-        return $text;
-    }
-}
+use Lotgd\SafeEscape;
+use Lotgd\Sanitize;
+use Lotgd\Stripslashes;
+use PHPUnit\Framework\TestCase;
 
 final class UtilityTest extends TestCase
 {
